Update same_stats implementation
Generalized the output shape in the three shapes: circles, lines or dots. The data for the shape are given via a csv file -> check example files to see the structure. Reduced csv read and write functions to what we need, see interface_CSV.ccp. Input and Output are now csv files.