calculation of distance matrix only once
distance matrix is being calculated only once and then the user can try multiple times with various parameters until the result is satifactory. Next step is to add the possibility of choosing multiple outputs in terms of frames. Further manage the parsing arguments, e.g. with the argp.h library to enable user help and options